Trukket lagring av endringer ut fra TilbakeTilOversiktLenke til en prop. Denne brukes ogsa utenfor AvtaleContext. Co-authored-by: Mats Bolstad <mats.bolstad@nav.no> Co-authored-by: Vegar Sechmann Molvig <vegar.sechmann.molvig@nav.no>
