+++ src/AvtaleSide/TilbakeTilOversiktLenke/TilbakeTilOversiktLenke.tsx
@@ -1,29 +1,16 @@
-import VenstreChevron from 'nav-frontend-chevron/lib/venstre-chevron';
-import React, { FunctionComponent, useContext } from 'react';
-import { Link } from 'react-router-dom';
 import { pathTilOversikt } from '@/paths';
-import { ApiError } from '@/types/errors';
-import { AvtaleContext } from '@/AvtaleContext';
+import VenstreChevron from 'nav-frontend-chevron/lib/venstre-chevron';
+import React, { FunctionComponent } from 'react';
+import { Link } from 'react-router-dom';
 import './TilbakeTilOversiktLenke.less';
 
-const TilbakeTilOversiktLenke: FunctionComponent<{}> = () => {
-    const context = useContext(AvtaleContext);
+type Props = {
+    onClick?: () => void;
+};
 
-    const tilbakeTilOversikt = async () => {
-        if (context.harUlagredeEndringer()) {
-            try {
-                await context.lagreAvtale();
-            } catch (error) {
-                if (error instanceof ApiError) {
-                    return context.visFeilmelding(error.message);
-                }
-                throw error;
-            }
-        }
-    };
-
+const TilbakeTilOversiktLenke: FunctionComponent<Props> = props => {
     return (
-        <Link to={pathTilOversikt} className="lenke" onClick={tilbakeTilOversikt}>
+        <Link to={pathTilOversikt} className="lenke" onClick={props.onClick}>
             <VenstreChevron className="tilbaketiloversikt__chevron" />
             Tilbake til oversikt
         </Link>
